refactor: split standard validation tech/era rules
